py/requirements.txt
gevent==22.10.2
greenlet==2.0.2
websocket-client==1.5.2
zope.event==4.6
zope.interface==6.0
